# swagger/pelican-swagger.yaml
swagger: "2.0"
info:
  title: Pelican Server APIs
  description:
    "[Pelican](https://pelicanplatform.org/) provides an open-source software platform for federating
    dataset repositories together and delivering the objects to computing capacity such as the [OSPool](https://osg-htc.org/services/open_science_pool.html)


    This is the API documentation for various APIs in Pelican servers (director, registry, origin, etc)
    to communicate with each other and in-between users accessing the servers.


    Note that we use cookie authentication and authorization. We check a cookie named `login` with value being a JWT.
    The cookie is issued after a successful call to `/api/v1.0/auth/login`. However, OpenAPI 2.0 does not support specifying cookie-based security check.
    Therefore, as an alternative, we will add `Authentication Required` to the API description where needed.

    For endpoint that is only accesible by users with admin privilege, we mark the endpoint as `Admin privilege Required`.


    For how to set up Pelican servers, please refer to the documentation at [docs.pelicanplatform.org](https://docs.pelicanplatform.org/)"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: API Support via Pelican GitHub Issue
    url: https://github.com/PelicanPlatform/pelican/issues
  version: "1.0"
basePath: /api/v1.0/
consumes:
  - application/json
produces:
  - application/json
schemes:
  - https
securityDefinitions:
  # This is a hacky way to specify Bearer token auth here because it's not actually an "apikey"
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      Enter the JWT with the `Bearer` prefix, e.g. "Bearer abcde12345".
definitions:
...
  ErrorModelV2:
    type: object
    description: The error response of a request
    properties:
      status:
        type: string
        default: "error"
        description: The response status
      msg:
        type: string
        default: ""
        description: The response message
...
  SuccessModelV2:
    type: object
    description: The success response of a request
    properties:
      status:
        type: string
        example: "success"
        description: The response status
      msg:
        type: string
        default: "success"
        description: The response message
...
  DirectorServerResponse:
    type: object
    properties:
      name:
        type: string
        example: "Test Origin"
      auth_url:
        type: string
        description: The URL of the issuer on the server to check token validity
        example: "https://example-origin.com:8443"
      broker_url:
        type: string
        description: The URL to the connection broker
        example: "https://example-origin.com:8447"
      url:
        type: string
        description: The URL of XrootD service on the server to access objects
        example: "https://example-origin.com:8443"
      web_url:
        type: string
        description: The URL of the web engine running on the server to visit the web interface and send API requests
        example: "https://example-origin.com:8444"
      type:
        type: string
        description: Type of the server. Origin|Cache
        example: Origin
      latitute:
        type: number
        description: The latitute of the server based on its IP address
        default: 0
      longitude:
        type: number
        description: The longitude of the server based on its IP address
        default: 0
      enableWrite:
        type: boolean
        description: If the origin server enables write access. Only valid for type=Origin
        default: false
      enableFallbackRead:
        type: boolean
        description: If fall back to read from the origin when there's no cache to serve the file
        default: false
      filtered:
        type: boolean
        description: |
          If the server is excluded from serving client object requests.
          No traffic will be redirect to the server if the value is true
        default: false
      filteredType:
        type: string
        description: |
          The type of filtering/allowance of the server. Can be permFilter|tempFilter|tempAllow

          * **permFilter** when the server is filtered by configuration parameter `Director.FilteredServers`
          * **tempFilter** when the server is filtered by web API request at `/servers/filter/*name`. The filter
            rule lives in-memory and will be reset at the server restart
          * **tempAllow** when the  server is filtered by configuration parameter but _allowed_ by the web API at `/servers/allow/*name`,
            the change lives in-memory and will be overwritten by config parameter at the server restart
        default: ""
      status:
        type: string
        description: The status of director file transfer test against the server. Can be Initializing|Unknown|OK|Error
        default: Unknown
      namespacePrefixes:
        type: array
        items:
          type: string
          minItems: 1
        default: []
        example: ["/foo", "/bar"]
        description: The namespaces the returned server provides
...
  DirectorContact:
    type: object
    properties:
      email:
        type: string
        default: ""
      url:
        type: string
        default: ""
...
paths:
...
  /servers:
    get:
      tags:
        - common
      summary: Returns a list of enabled servers
      description: Server names are in lower-case, sorted in alphabetical order.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              servers:
                type: array
                items:
                  type: string
                  minItems: 1
                example: ["director", "origin", "registry"]
        "500":
          description: Server encountered error in reading institution configuration
          schema:
            type: object
            $ref: "#/definitions/ErrorModelV2"
...
  /director_ui/servers:
    get:
      tags:
        - "director_ui"
      summary: Get a list of storage servers advertised to the director
      produces:
        - application/json
      parameters:
        - in: query
          name: server_type
          type: string
          description: The server type to filter the list. Can be origin|cache
      responses:
        "200":
          description: "OK"
          schema:
            type: array
            items:
              type: object
              $ref: "#/definitions/DirectorServerResponse"
              minimum: 0
        "400":
          description: "Bad request, query parameter is invalid"
          schema:
            type: object
            $ref: "#/definitions/ErrorModelV2"
    patch:
      summary: Toggle disable/enable a server from accepting object transfer requests
      description: |
        `Authentication Required` `Admin privilege Required`

        **Requests with the body `{disabled: false }` on an already-disabled server are in-memory and will be removed at the server restart.**

        You may reset the rules defined in `Director.DisabledServers` in-memory but they will be overwritten at the server restart
      tags:
        - "director_ui"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: query
          name: serverUrl
          type: string
          required: true
          description: |
            The **URL-encoded** server URL to disable/enable the server.
            The value is the `url` field in `GET /director/servers` response.
        - in: body
          name: server
          required: true
          schema:
            type: object
            properties:
              disabled:
                type: boolean
      responses:
        "200":
          description: "OK"
          schema:
            type: object
            $ref: "#/definitions/SuccessModelV2"
        "400":
          description: "Bad request. Either `serverUrl` is missing or the request body is invalid"
          schema:
            type: object
            $ref: "#/definitions/ErrorModelV2"
        "403":
          description: "Forbidden. Admin privilege required"
          schema:
            type: object
            $ref: "#/definitions/ErrorModelV2"
...
  /director_ui/contact:
    get:
      summary: Get the support contact information of the federation the director hostnames
      description: "Note that the response value (if successful) `email` and `url` can be either or both empty strings"
      tags:
        - "director_ui"
      responses:
        "200":
          description: "OK"
          schema:
            type: object
            $ref: "#/definitions/DirectorContact"
        "500":
          description: "Bad server configuration. Failed to get the director contact information from the server configuration"
          schema:
            type: object
            $ref: "#/definitions/ErrorModelV2"
...